Feature: Implementación sugerencias en asesoria con tutor Se ajusta double  a big decimal Implementación response codes Ajuste en caso de uso para validacion y codigos para BusinessException Estandarizacion de respuesta con ApiResponse Implementación de TechnicalException Creación de constantes para string quemados

+++ domain/model/src/main/java/co/com/bancolombia/model/user/User.java
@@ -1,6 +1,8 @@
 package co.com.bancolombia.model.user;
 
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -19,5 +21,5 @@ public class User {
     private String phone;
     private String email;
     private Integer idRole;
-    private Double baseSalary;
+    private BigDecimal baseSalary;
 }
